Preemptive useless commit
Well, actually, I was going to be doing the provisioning of the infrastructure
in this here commit, but um... decided against it, so the name of the branch is
.. pointless. So, instead, it's now just basically documentation updates, to
which someone might consider using to setup their workstation/environment for
developing the code used in the Azure Function App.

The infrastructure as code was moved to unlettered-infra. Super name for it.